fix(warn): 编译警告清理——修复 10 处真问题/可读性问题
- OperatorPrecedence(JavaCodegen.isHex)：混合 &&/|| 改局部布尔避免优先级歧义（checkstyle 禁多余括号，用拆分变量替代）
- MissingOverride(AbstractJavaGenerator.generate)、MixedMutability(EnumCommentParser 返回不可变)
- IntLongMath(RoundTripSmokeTest.hexValue 长运算)、InvalidBlockTag×2（EndToEnd/GenerateMojoTest 误 javadoc @TempDir）
- UnusedVariable×2 内部参数（GenerationContext.syntheticTable / AbstractJavaGenerator.buildFresh）
门禁全绿；警告 52→42（余 42 为 checker 注解风格 30 + StringSplitter 6 + 覆写/测试参数 3 + 命名冲突 2 + 三方 deprecation 1）

+++ ddl-codegen-tree/src/test/java/hyc/codegen/tree/RoundTripSmokeTest.java
@@ -31,9 +31,9 @@ public class RoundTripSmokeTest {
 
     private static long hexValue(char c) {
         if (c >= '0' && c <= '9') {
-            return c - '0';
+            return (long)c - '0';
         }
-        return (c >= 'a' && c <= 'f') ? (c - 'a' + 10) : (c - 'A' + 10);
+        return (c >= 'a' && c <= 'f') ? ((long)c - 'a' + 10) : ((long)c - 'A' + 10);
     }
 
     private static boolean isHexDigit(char c) {
